# backend/docs/openapi-exportai.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ExportAI API",
    "description": "API do ExportAI para recomendacao personalizada de mercados internacionais por NCM ou HS6.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/diagnosticos": {
      "post": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Cria uma sess�o de diagn�stico",
        "operationId": "criar_diagnostico_api_v1_diagnosticos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarDiagnosticoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoCriadaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosticos/{diagnostico_id}": {
      "get": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Consulta uma sess�o de diagn�stico",
        "operationId": "consultar_diagnostico_api_v1_diagnosticos__diagnostico_id__get",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoDiagnosticoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosticos/{diagnostico_id}/mercados": {
      "put": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Atualiza os pa�ses escolhidos",
        "operationId": "atualizar_mercados_api_v1_diagnosticos__diagnostico_id__mercados_put",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarMercadosRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MercadosAtualizadosResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosticos/{diagnostico_id}/respostas": {
      "put": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Registra respostas parciais ou completas",
        "operationId": "registrar_respostas_api_v1_diagnosticos__diagnostico_id__respostas_put",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrarRespostasRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostasRegistradasResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosticos/{diagnostico_id}/calcular": {
      "post": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Calcula o resultado do diagn�stico",
        "operationId": "calcular_diagnostico_api_v1_diagnosticos__diagnostico_id__calcular_post",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoDiagnostico"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosticos/{diagnostico_id}/resultado": {
      "get": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Consulta o �ltimo resultado calculado",
        "operationId": "consultar_resultado_api_v1_diagnosticos__diagnostico_id__resultado_get",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoDiagnostico"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosticos/{diagnostico_id}/pesquisa-assistida": {
      "post": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Executa a interpreta��o complementar por pa�s",
        "operationId": "executar_pesquisa_assistida_api_v1_diagnosticos__diagnostico_id__pesquisa_assistida_post",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutarPesquisaAssistidaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoPesquisaAssistida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Consulta a �ltima pesquisa assistida",
        "operationId": "consultar_pesquisa_assistida_api_v1_diagnosticos__diagnostico_id__pesquisa_assistida_get",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoPesquisaAssistida"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/diagnosticos/{diagnostico_id}/pesquisa-assistida/automatica": {
      "post": {
        "tags": [
          "Diagn�stico"
        ],
        "summary": "Busca fontes autorizadas e executa interpreta��o complementar",
        "operationId": "executar_pesquisa_assistida_automatica_api_v1_diagnosticos__diagnostico_id__pesquisa_assistida_automatica_post",
        "parameters": [
          {
            "name": "diagnostico_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Diagnostico Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutarPesquisaAutomaticaRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoPesquisaAutomatica"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jornadas/{journey_id}/diagnostico": {
      "put": {
        "tags": [
          "Jornadas an�nimas"
        ],
        "summary": "Associar Diagnostico",
        "operationId": "associar_diagnostico_api_v1_jornadas__journey_id__diagnostico_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "journey_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Journey Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociarDiagnosticoRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JornadaPublica"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AtualizarMercadosRequest": {
        "properties": {
          "mercados_escolhidos": {
            "items": {
              "type": "string",
              "pattern": "^[A-Z]{3}$"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Mercados Escolhidos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "mercados_escolhidos"
        ],
        "title": "AtualizarMercadosRequest"
      },
...
      "CriarDiagnosticoRequest": {
        "properties": {
          "produto": {
            "$ref": "#/components/schemas/ProdutoDiagnostico"
          },
          "mercados_basico": {
            "items": {
              "$ref": "#/components/schemas/MercadoBasico"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Mercados Basico"
          },
          "mercados_escolhidos": {
            "items": {
              "type": "string",
              "pattern": "^[A-Z]{3}$"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Mercados Escolhidos"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "produto",
          "mercados_basico",
          "mercados_escolhidos"
        ],
        "title": "CriarDiagnosticoRequest"
      },
...
      "ExecutarPesquisaAssistidaRequest": {
        "properties": {
          "documentos_por_pais": {
            "items": {
              "$ref": "#/components/schemas/DocumentosPorPais"
            },
            "type": "array",
            "maxItems": 3,
            "title": "Documentos Por Pais"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExecutarPesquisaAssistidaRequest"
      },
      "ExecutarPesquisaAutomaticaRequest": {
        "properties": {
          "incluir_documentos_manuais": {
            "type": "boolean",
            "title": "Incluir Documentos Manuais",
            "default": true
          },
          "maximo_resultados_por_consulta": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Maximo Resultados Por Consulta",
            "default": 3
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ExecutarPesquisaAutomaticaRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MercadosAtualizadosResponse": {
        "properties": {
          "diagnostico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Diagnostico Id"
          },
          "mercados_escolhidos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mercados Escolhidos"
          },
          "resultado_invalidado": {
            "type": "boolean",
            "title": "Resultado Invalidado"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "diagnostico_id",
          "mercados_escolhidos",
          "resultado_invalidado"
        ],
        "title": "MercadosAtualizadosResponse"
      },
...
      "RegistrarRespostasRequest": {
        "properties": {
          "catalogo_versao": {
            "type": "string",
            "title": "Catalogo Versao"
          },
          "respostas": {
            "items": {
              "$ref": "#/components/schemas/RespostaChecklist"
            },
            "type": "array",
            "minItems": 1,
            "title": "Respostas"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "catalogo_versao",
          "respostas"
        ],
        "title": "RegistrarRespostasRequest"
      },
...
      "RespostasRegistradasResponse": {
        "properties": {
          "diagnostico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Diagnostico Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "CRIADO",
              "EM_PREENCHIMENTO",
              "PRONTO_PARA_CALCULAR",
              "CALCULADO"
            ],
            "title": "Status"
          },
          "respostas_registradas": {
            "type": "integer",
            "title": "Respostas Registradas"
          },
          "total_itens_ativos": {
            "type": "integer",
            "title": "Total Itens Ativos"
          },
          "faltam_responder": {
            "type": "integer",
            "title": "Faltam Responder"
          },
          "resultado_invalidado": {
            "type": "boolean",
            "title": "Resultado Invalidado"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "diagnostico_id",
          "status",
          "respostas_registradas",
          "total_itens_ativos",
          "faltam_responder",
          "resultado_invalidado"
        ],
        "title": "RespostasRegistradasResponse"
      },
...
      "ResultadoPesquisaAutomatica": {
        "properties": {
          "busca_provedor": {
            "type": "string",
            "title": "Busca Provedor"
          },
          "estatisticas": {
            "items": {
              "$ref": "#/components/schemas/EstatisticaBuscaPais"
            },
            "type": "array",
            "title": "Estatisticas"
          },
          "pesquisa": {
            "$ref": "#/components/schemas/ResultadoPesquisaAssistida"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "busca_provedor",
          "estatisticas",
          "pesquisa"
        ],
        "title": "ResultadoPesquisaAutomatica"
      },
...
      "SessaoCriadaResponse": {
        "properties": {
          "diagnostico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Diagnostico Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "CRIADO",
              "EM_PREENCHIMENTO",
              "PRONTO_PARA_CALCULAR",
              "CALCULADO"
            ],
            "title": "Status"
          },
          "catalogo_versao": {
            "type": "string",
            "title": "Catalogo Versao"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "diagnostico_id",
          "status",
          "catalogo_versao",
          "criado_em"
        ],
        "title": "SessaoCriadaResponse"
      },
      "SessaoDiagnosticoResponse": {
        "properties": {
          "diagnostico_id": {
            "type": "string",
            "format": "uuid",
            "title": "Diagnostico Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "CRIADO",
              "EM_PREENCHIMENTO",
              "PRONTO_PARA_CALCULAR",
              "CALCULADO"
            ],
            "title": "Status"
          },
          "catalogo_versao": {
            "type": "string",
            "title": "Catalogo Versao"
          },
          "produto": {
            "$ref": "#/components/schemas/ProdutoDiagnostico"
          },
          "mercados_basico": {
            "items": {
              "$ref": "#/components/schemas/MercadoBasico"
            },
            "type": "array",
            "title": "Mercados Basico"
          },
          "mercados_escolhidos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Mercados Escolhidos"
          },
          "respostas": {
            "items": {
              "$ref": "#/components/schemas/RespostaChecklist"
            },
            "type": "array",
            "title": "Respostas"
          },
          "resultado": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultadoDiagnostico"
              },
              {
                "type": "null"
              }
            ]
          },
          "pesquisa_assistida": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pesquisa Assistida"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "diagnostico_id",
          "status",
          "catalogo_versao",
          "produto",
          "mercados_basico",
          "mercados_escolhidos",
          "respostas",
          "resultado",
          "criado_em",
          "atualizado_em"
        ],
        "title": "SessaoDiagnosticoResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}